there is this very cool extension that allows to do this. In general, i've learned that it is much safer to pattern match this way (with or without this extension) versus using accessors because accessors are sometimes incomplete functions. Consider:
data Foo = Bar{ stuff :: Int} | Baz

if you do
  f :: Foo -> ...
  f x = stuff x
and x was not constructed with Bar, the program will fail at runtime.

this version however is safer:
  f Bar{stuff} = ... do something with stuff ...
but of course the incompleteness is shifted to f. At least this is more
explicit.

Typechecker warns you about incomplete functions however, which brings
me to another suggestion - use the -Werror flag in your cabal file and
get cleanup all the warnings